tests/wpt/web-platform-tests/xhr/send-non-same-origin.htm

<!DOCTYPE html>
<html>
  <head>
    <title>XMLHttpRequest: send() - non same-origin</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <base>
  </head>
  <body>
    <div id="log"></div>
    <script src="/common/get-host-info.sub.js"></script>
    <script>
      // Setting base URL before running the tests
      var host_info = get_host_info();
      document.getElementsByTagName("base")[0].setAttribute("href", host_info.HTTP_REMOTE_ORIGIN);

      function url(url) {
        test(function() {
          var client = new XMLHttpRequest()
          client.open("GET", url, false)
          assert_throws_dom("NetworkError", function() { client.send() })
        }, document.title + " (" + url + ")")
      }
      url("mailto:test@example.org")
      url("tel:+31600000000")
      url(host_info.HTTP_REMOTE_ORIGIN)
      url("javascript:alert('FAIL')")
      url("folder.txt")
      url("about:blank")
      url("blob:bogusidentifier")
    </script>
  </body>
</html>
